+++ src/Orchestra/Memory/Console/MemoryCommand.php
@@ -11,7 +11,7 @@ class MemoryCommand extends Command
      *
      * @var string
      */
-    protected $name = 'orchestra:memory';
+    protected $name = 'memory:migrate';
 
     /**
      * The console command description.
@@ -21,61 +21,12 @@ class MemoryCommand extends Command
     protected $description = 'Orchestra\Memory Command';
 
     /**
-     * Create a new command instance.
-     *
-     * @return void
-     */
-    public function __construct()
-    {
-        parent::__construct();
-    }
-
-    /**
      * Execute the console command.
      *
      * @return void
      */
     public function fire()
     {
-        $action = $this->argument('action');
-
-        if (in_array($action, array('install', 'upgrade'))) {
-            $this->fireMigration();
-            $this->info('orchestra/memory has been migrated');
-        } else {
-            $this->error("Invalid action [{$action}].");
-        }
-    }
-
-    /**
-     * Fire migration process.
-     *
-     * @return void
-     */
-    protected function fireMigration()
-    {
         $this->call('migrate', array('--package' => 'orchestra/memory'));
-    }
-
-    /**
-     * Get the console command arguments.
-     *
-     * @return array
-     */
-    protected function getArguments()
-    {
-        return array(
-            array('action', InputArgument::REQUIRED, "Type of action, e.g: 'install', 'upgrade'."),
-        );
-    }
-
-    /**
-     * Get the console command options.
-     *
-     * @return array
-     */
-    protected function getOptions()
-    {
-        return array();
     }
 }