+++ docs/changes.md
@@ -7,6 +7,7 @@ Memory Change Log
 
 * Add `Orchestra\Memory\Abstractable\Container`.
 * Predefined package path to avoid additional overhead to guest package path.
+* Rename command to `php artisan memory:migrate`.
 
 ## Version 2.0